pyproject.toml
[tool.poetry]
name = "aqimon"
version = "0.1.9"
description = "Air Quality Index Monitor"
authors = ["Tim Orme"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.92.0"
pyserial = "^3.5"
uvicorn = "^0.21.0"
databases = {extras = ["aiosqlite"], version = "^0.7.0"}
jinja2 = "^3.1.2"
fastapi-utils = "^0.2.1"
pyserde = {extras = ["toml"], version = "^0.10.2"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.256"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
aqimon = "aqimon.server:start"
aqimond = "aqimon.server:debug"

[tool.ruff]
select = ["E", "F", "D"]
ignore = ["D203", "D213"]
line-length = 120

[tool.black]
line-length = 120